.history/requirements_20200712115541.txt
django
pytest
mixer
pytest-django
pytest-cov
djangorestframework